+++ examples/example_pygame.py
@@ -23,24 +23,13 @@ def main():
         screen.fill(BLACK)
 
         with cui.CoshUIRenderer(cui.PygameBackend(screen), cui.DEBUG):
-            with cui.Container(id="container_1", width=cui.FILL, height=cui.FILL, style=cui.CoshStyling(background_color=(80, 75, 255)), align=cui.ALIGN_CENTER, justify=cui.JUSTIFY_CENTER):
-                with cui.Container(id="main_container", direction=cui.COLUMN, align=cui.ALIGN_CENTER, justify=cui.JUSTIFY_CENTER, gap=15, style=cui.CoshStyling(background_color=(255, 200, 200))):
-                    cui.Label(id="main_label", text="CoshUI Menu", font_size=52)
-                    cui.Button(id="settings_button", text="Settings is a stupid thing to talk about and I don't like it", text_overflow=cui.TEXT_WRAP, height=50, text_justify=cui.TEXT_JUSTIFY_CENTER)
-                    cui.Button(id="quit_button", text="Quit", height=50, width=150, style=cui.CoshStyling(transform_rotation=45.0))
-                    cui.Image(id="test_image", src="assets/image.png", width=75, height=75, style=cui.CoshStyling(transform_rotation=45.0))
+            with cui.Container(id="main_root", width=cui.FILL, height=cui.FILL, padding=10):
+                cui.Container(id="test", width=100, height=100, style=cui.CoshStyling(background_color=(255, 255, 0)))
 
-        if cui.get_signal("settings_button", cui.CLICKED):
-            cui.animate("transform_scale", "main_label", 1.2, 0.25, "ease_in")
-            cui.animate("background_color", "main_container", (255, 0, 0), 1.5, "ease_in")
-        if cui.get_signal("settings_button", cui.RELEASED):
-            cui.animate("transform_scale", "main_label", 1.0, 0.25, "ease_in")
-            cui.animate("background_color", "main_container", (255, 200, 200), 1.5, "ease_out")
-
-        if cui.get_signal("quit_button", cui.HOVERED):
-            cui.animate("transform_rotation", "quit_button", 0.0, 0.5, "ease_in")
-        if cui.get_signal("quit_button", cui.HOVER_EXIT):
-            cui.animate("transform_rotation", "quit_button", 45.0, 0.5, "ease_in")
+        if cui.get_signal("test", cui.CLICKED):
+            cui.animate("transform_rotation", "test", 45.0, 1.0, "ease_out_bounce")
+        if cui.get_signal("test", cui.RELEASED):
+            cui.animate("transform_rotation", "test", 0.0, 1.0, "ease_out_bounce")
         if cui.get_signal("quit_button", cui.CLICKED):
             running = False