+++ src/coshui/backends/graphics/gl_window_drivers.py
@@ -15,7 +15,7 @@ except ImportError:
 
 class _WindowDriver(ABC):
     @abstractmethod
-    def _measure_text(self):
+    def _measure_text(self, text, font_path, font_size):
         pass
     
     @abstractmethod
@@ -24,10 +24,6 @@ class _WindowDriver(ABC):
     
     @abstractmethod
     def _get_size(self):
-        pass
-
-    @abstractmethod
-    def _get_ortho_projection(self):
         pass
 
 # GLFW
@@ -43,22 +39,21 @@ class _GLFWDriver(_WindowDriver):
             CoshInput._mouse_position[1] - CoshInput._prev_mouse_position[1]
         )
         CoshInput._prev_mouse_position = CoshInput._mouse_position
-        CoshInput._current_mouse_pressed = True if glfw.get_mouse_button(glfw.get_current_context(), glfw.MOUSE_BUTTON_LEFT) is glfw.PRESS else False
+        CoshInput._current_mouse_pressed = glfw.get_mouse_button(glfw.get_current_context(), glfw.MOUSE_BUTTON_LEFT) is glfw.PRESS
 
     def _get_size(self):
         return glfw.get_window_size(glfw.get_current_context())
-
-    def _get_ortho_projection(self):
-        pass
 
 # MGLW
 class _MGLWDriver(_WindowDriver):
     def _measure_text(self):
         pass
 
+    # This works only if _calc_mouse_delta(self.mouse_x, self.mouse_y) is called beforehand
+    # or if _mouse_pos is set somewhere. Either way, MGLW hates others making tooling around it seems.
     def _poll_input(self):
         CoshInput._prev_mouse_pressed = CoshInput._current_mouse_pressed
-        CoshInput._mouse_position = moderngl_window.window()._mouse_pos # Doesn't seem to work.
+        CoshInput._mouse_position = moderngl_window.window()._mouse_pos
         CoshInput._mouse_delta = (
             CoshInput._mouse_position[0] - CoshInput._prev_mouse_position[0],
             CoshInput._mouse_position[1] - CoshInput._prev_mouse_position[1]
@@ -68,10 +63,6 @@ class _MGLWDriver(_WindowDriver):
 
     def _get_size(self):
         return moderngl_window.window().size
-
-    def _get_ortho_projection(self):
-        width, height = self._get_size()
-        return moderngl_window.geometry.projection.ortho(0, width, height, 0, -1, 1)
 
 class Windower(Enum):
     GLFW = _GLFWDriver